Split up tests into multiple modules, implement scikit-learn estimator checks on non-subclassed estimators, remove redundant tests and test parametrizations.
