+++ .github/workflows/tests.yaml
@@ -65,7 +65,11 @@ jobs:
       - name: Test with pytest 
         run: |
           poetry show
-          poetry run python -m pytest -v --cov=scikeras --cov-report xml --numprocesses=auto
+          poetry run python -m pytest --cov=scikeras --cov-report xml --numprocesses=auto
+      
+      - name: DEBUGGING - Run tests sequentially
+        if: ! ${{ success() }}
+        run: poetry run python -m pytest -v
 
       - uses: codecov/codecov-action@v1
 
@@ -123,7 +127,11 @@ jobs:
       - name: Test with pytest 
         run: |
           poetry show
-          poetry run python -m pytest -v --cov=scikeras --cov-report xml --numprocesses=auto
+          poetry run python -m pytest --cov=scikeras --cov-report xml --numprocesses=auto
+      
+      - name: DEBUGGING - Run tests sequentially
+        if: ! ${{ success() }}
+        run: poetry run python -m pytest -v
 
       - uses: codecov/codecov-action@v1
 
@@ -172,7 +180,11 @@ jobs:
       - name: Test with pytest 
         run: |
           poetry show
-          poetry run python -m pytest -v --cov=scikeras --cov-report xml --numprocesses=auto
+          poetry run python -m pytest --cov=scikeras --cov-report xml --numprocesses=auto
+      
+      - name: DEBUGGING - Run tests sequentially
+        if: ! ${{ success() }}
+        run: poetry run python -m pytest -v
 
       - uses: codecov/codecov-action@v1
 
@@ -217,6 +229,10 @@ jobs:
       - name: Test with pytest
         run: |
           poetry show
-          poetry run python -m pytest -v --cov=scikeras --cov-report xml --numprocesses=auto
+          poetry run python -m pytest --cov=scikeras --cov-report xml --numprocesses=auto
+      
+      - name: DEBUGGING - Run tests sequentially
+        if: ! ${{ success() }}
+        run: poetry run python -m pytest -v
 
       - uses: codecov/codecov-action@v1
